refactor: Add admin_required decorator for authorization
This commit adds a new decorator `admin_required` to the `utils.py` file. The decorator checks if the current user has admin privileges before executing the decorated function. If the user is not an admin, a 403 Forbidden error is raised. This decorator can be used to protect routes or functions that require admin access.