refactor: Initialise controllers in controller init
This commit adds the controller blueprints to the `__init__.py` file from the  individual files for better organization and maintainability by reducing the likeliness of a circular import. Each controller now has its definition in the `__init__.py` file. This refactoring improves code readability and makes it easier to manage and extend the application's functionality.