feat: Paginate comments in get_comments endpoint
This commit modifies the `get_comments` endpoint in the `comment_controller.py` file to implement pagination for retrieving comments. The `page` and `per_page` query parameters are added to allow users to navigate through the comments. This enhancement improves the performance and user experience when dealing with a large number of comments.