feat: Add comments_count field to PostSchema
This commit adds the `comments_count` field to the `PostSchema` in the `post.py` file. The `comments_count` field is a method that calculates the number of comments a post has. This enhancement improves the functionality of the application by providing the ability to display the number of comments for each post.