feat: Add JWT authentication to get_users route
This commit adds JWT authentication to the `get_users` route in the `user_controller.py` file. By using the `@jwt_required()` decorator, only authenticated users will be able to access the list of users. This enhancement improves the security of the application and ensures that sensitive user information is protected.