refactor: Add delete user endpoint
This commit adds a new endpoint `/unregister` to the `auth_controller.py` file. The endpoint is responsible for deleting a user from the database. It requires authentication using JWT and retrieves the user ID from the token. The user is then fetched from the database and deleted. Finally, a JSON message indicating the status of the deletion is returned. This change improves the functionality of the authentication system.